farukon_core::data_handler

Trait DataHandler

pub trait DataHandler {
    // Required methods
    fn get_latest_bar(&self, symbol: &str) -> Option<SOAData>;
    fn get_latest_bars(&self, symbol: &str, n: usize) -> Option<SOAData>;
    fn get_latest_bar_datetime(&self, symbol: &str) -> Option<DateTime<Utc>>;
    fn get_latest_bar_value(&self, symbol: &str, val_type: &str) -> Option<f64>;
    fn get_latest_bars_values(
        &self,
        symbol: &str,
        val_type: &str,
        n: usize,
    ) -> Option<&[f64]>;
    fn get_latest_bars_volume_values(
        &self,
        symbol: &str,
        n: usize,
    ) -> Option<&[u64]>;
    fn update_bars(&mut self);
    fn get_continue_backtest(&self) -> bool;
    fn set_continue_backtest(&mut self, value: bool);
}

Defines the interface for a data handler. All data handlers must implement this trait.

Required Methods

fn get_latest_bar(&self, symbol: &str) -> Option<SOAData>

Returns the latest bar for the specified symbol.

Arguments

symbol - The symbol to retrieve data for.

Returns

An optional reference to the latest MarketBar.

fn get_latest_bars(&self, symbol: &str, n: usize) -> Option<SOAData>

Returns the last n bars for the specified symbol.

Arguments

symbol - The symbol to retrieve data for.

n - The number of bars to retrieve.

Returns

A vector of references to the last n MarketBars.

fn get_latest_bar_datetime(&self, symbol: &str) -> Option<DateTime<Utc>>

Returns the timestamp of the latest bar for the specified symbol.

Arguments

symbol - The symbol to retrieve data for.

Returns

An optional chrono::DateTime<chrono::Utc> representing the timestamp.

fn get_latest_bar_value(&self, symbol: &str, val_type: &str) -> Option<f64>

Returns a specific value (open, high, low, close, volume) from the latest bar for the specified symbol.

Arguments

symbol - The symbol to retrieve data for.

val_type - The type of value to retrieve (“open”, “high”, “low”, “close”, “volume”).

Returns

An optional f64 representing the value, or None if not available.

fn get_latest_bars_values( &self, symbol: &str, val_type: &str, n: usize, ) -> Option<&[f64]>

Returns a specific value (open, high, low, close, volume) from the last n bars for the specified symbol.

Arguments

symbol - The symbol to retrieve data for.

val_type - The type of value to retrieve (“open”, “high”, “low”, “close”, “volume”).

n - The number of bars to retrieve.

Returns

A vector of f64 values.

fn get_latest_bars_volume_values( &self, symbol: &str, n: usize, ) -> Option<&[u64]>

fn update_bars(&mut self)

Advances the data handler to the next bar. This method is called by the backtesting engine to simulate time passing.

fn get_continue_backtest(&self) -> bool

Advances the data handler to the next bar. This method is called by the backtesting engine to simulate time passing.

fn set_continue_backtest(&mut self, value: bool)

Sets the flag to stop the backtest.

Arguments

value - The new value for the flag.