farukon_core

Module settings

Configuration structures for the Farukon platform. Loads settings from JSON files and validates them.

Structs

CommonSettings

Common settings applicable to the entire platform.

DataSettings

Data settings for a strategy.

FitnessParams

Fitness function parameters for the Genetic Algorithm.

GAParams

Parameters for the Genetic Algorithm optimizer.

MarginParams

PortfolioSettingsForStrategy

Portfolio settings for a strategy.

PosSizer

Position sizing parameters.

Settings

Top-level settings structure.

StrategySettings

Settings for a single strategy.

Enums

FitnessValue

Type of fitness metric to optimize.

KellyMode

MetricsMode

Mode for calculating performance metrics.

OptimizerType

Type of optimizer to use.

ParamSpec

Represents a parameter specification that can be either a discrete list of values or a continuous range.