farukon_core::utils

Function generate_lhs_parameter_sets

pub fn generate_lhs_parameter_sets(
    strategy_params_ranges: &HashMap<String, ParamSpec>,
    pos_sizer_value_range: &ParamSpec,
    slippage_range: &ParamSpec,
    pos_sizer_additional_params: &HashMap<String, ParamSpec>,
    pos_sizer_name: &String,
    size: usize,
    rng: &mut impl Rng,
) -> Vec<ParameterSet>

Generates a vector of ParameterSet objects using Latin Hypercube Sampling (LHS) for specified parameter ranges.

This function creates a diverse initial population or fresh blood by sampling parameter values according to their specifications (either continuous Range or discrete Discrete lists). It ensures that the generated sets cover the defined parameter space effectively.

Arguments

strategy_params_ranges - A map of strategy parameter names to their ParamSpec.

pos_sizer_value_range - The ParamSpec for the position sizer value.

slippage_range - The ParamSpec for the slippage value.

pos_sizer_additional_params - A map of additional position sizer parameter names to their ParamSpec.

pos_sizer_name - The name of the position sizing method.

size - The number of ParameterSet objects to generate.

rng - A mutable reference to a random number generator implementing rand::Rng.

Returns

A Vec<optimization::ParameterSet> containing the generated parameter sets.